Move all MIME types to a single module...
and added an HTML MIME type
