+++ webgear-core/src/WebGear/Core.hs
@@ -22,17 +22,13 @@ module WebGear.Core (
   module WebGear.Core.Handler,
   module WebGear.Core.Handler.Static,
   module WebGear.Core.Traits,
-  module WebGear.Core.MIMEType,
-  module WebGear.Core.MIMEType.JSON,
-  module WebGear.Core.MIMEType.OctetStream,
-  module WebGear.Core.MIMEType.PlainText,
-  module WebGear.Core.MIMEType.FormURLEncoded,
-  module WebGear.Core.MIMEType.MultiPartFormData,
+  module WebGear.Core.MIMETypes,
 ) where
 
 import Control.Arrow
 import Data.Text
 import qualified Network.Wai as Wai
+import WebGear.Core.MIMETypes
 
 import WebGear.Core.Handler
 import WebGear.Core.Handler.Static
@@ -41,10 +37,3 @@ import WebGear.Core.Request
 import WebGear.Core.Response
 import WebGear.Core.Trait
 import WebGear.Core.Traits
-
-import WebGear.Core.MIMEType
-import WebGear.Core.MIMEType.FormURLEncoded
-import WebGear.Core.MIMEType.JSON
-import WebGear.Core.MIMEType.MultiPartFormData
-import WebGear.Core.MIMEType.OctetStream
-import WebGear.Core.MIMEType.PlainText
