+++ webgear-core/src/WebGear/Core/Trait/Auth/Common.hs
@@ -21,7 +21,7 @@ import Data.Void (absurd)
 import GHC.TypeLits (KnownSymbol, Symbol, symbolVal)
 import Web.HttpApiData (FromHttpApiData (..))
 import WebGear.Core.Handler (Handler, unwitnessA, (>->))
-import WebGear.Core.MIMEType.PlainText (PlainText (..))
+import WebGear.Core.MIMETypes (PlainText (..))
 import WebGear.Core.Modifiers (Existence (..), ParseStyle (..))
 import WebGear.Core.Request (Request)
 import WebGear.Core.Response (Response)
