+++ webgear-core/webgear-core.cabal
@@ -92,12 +92,7 @@ library
                     , WebGear.Core.Request
                     , WebGear.Core.Response
                     , WebGear.Core.Handler
-                    , WebGear.Core.MIMEType
-                    , WebGear.Core.MIMEType.FormURLEncoded
-                    , WebGear.Core.MIMEType.JSON
-                    , WebGear.Core.MIMEType.MultiPartFormData
-                    , WebGear.Core.MIMEType.OctetStream
-                    , WebGear.Core.MIMEType.PlainText
+                    , WebGear.Core.MIMETypes
                     , WebGear.Core.Traits
                     , WebGear.Core.Trait.Auth.Basic
                     , WebGear.Core.Trait.Auth.JWT
