+++ webgear-openapi/src/WebGear/OpenApi/Trait/Body.hs
@@ -9,7 +9,7 @@ import Data.OpenApi.Declare (runDeclare)
 import Data.Proxy (Proxy (..))
 import Data.Text (Text)
 import GHC.Exts (fromList)
-import WebGear.Core.MIMEType (MIMEType (..))
+import WebGear.Core.MIMETypes (MIMEType (..))
 import WebGear.Core.Request (Request)
 import WebGear.Core.Response (Response (..), ResponseBody)
 import WebGear.Core.Trait (Get (..), Set (..), With)
