+++ webgear-swagger/src/WebGear/Swagger/Trait/Body.hs
@@ -8,7 +8,7 @@ import Data.Proxy (Proxy (..))
 import Data.Swagger hiding (Response)
 import Data.Swagger.Declare (runDeclare)
 import Data.Text (Text)
-import WebGear.Core.MIMEType (MIMEType (..))
+import WebGear.Core.MIMETypes (MIMEType (..))
 import WebGear.Core.Request (Request)
 import WebGear.Core.Response (Response (..), ResponseBody)
 import WebGear.Core.Trait (Get (..), Set (..), With)
